pyproject.toml
[tool.poetry]
name = "educational_platform"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.6"
python-dotenv = "1.0.0"
psycopg2 = "^2.9.9"
pre-commit = "^3.5.0"
pillow = "^10.1.0"
django-ckeditor = "^6.7.0"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.1.0"
pylint = "^3.0.2"
black = "^23.10.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
